Move delete button to profile edit page
