+++ src/pages/Profile/ProfileUtilities.tsx
@@ -276,16 +276,6 @@ export const useProfileDisplayStyling = () => {
         width: '9rem',
     };
 
-    const regDeleteButton: React.CSSProperties = {
-        ...button,
-        width: '9rem',
-        background: '#ff0000',
-    };
-
-    const regDeleteButtonHover: React.CSSProperties = {
-        backgroundColor: '#ff3f3f',
-    };
-
     return {
         regProfileContainer,
         regProfileInfoContainer,
@@ -298,8 +288,6 @@ export const useProfileDisplayStyling = () => {
         regProfileFieldValue,
         regButtonContainer,
         regProfileButton,
-        regDeleteButton,
-        regDeleteButtonHover,
     };
 };
 
@@ -396,6 +384,16 @@ export const useEditProfileStyling = () => {
         
     };
 
+    const regDeleteButton: React.CSSProperties = {
+        ...button,
+        width: '9rem',
+        background: '#ff0000',
+    };
+
+    const regDeleteButtonHover: React.CSSProperties = {
+        backgroundColor: '#ff3f3f',
+    };
+
     return {
         editProfileContainer,
         editProfileHeader,
@@ -405,6 +403,8 @@ export const useEditProfileStyling = () => {
         profileImageEditor,
         profilePictureContainer,
         uploadImagebutton,
+        regDeleteButton,
+        regDeleteButtonHover,
     };
 };
